+++ projects/ngx-libs/src/lib/ngx-shade/index.ts
@@ -1,0 +1,2 @@
+export { NgxShadeComponent } from './ngx-shade.component';
+export { NgxShadeModule } from './ngx-shade.module';