+++ projects/ngx-libs/src/lib/ngx-shade/ngx-shade.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { NgxShadeComponent } from './ngx-shade.component';
+
+@NgModule({
+  imports: [
+    CommonModule
+  ],
+  declarations: [NgxShadeComponent],
+  exports: [NgxShadeComponent]
+})
+export class NgxShadeModule { }