+++ projects/ngx-libs/src/public_api.ts
@@ -3,3 +3,4 @@
  */
 
 export * from './lib/ngx-libs.module';
+export * from './lib/ngx-shade';