+++ src/app/app.module.ts
@@ -3,11 +3,13 @@ import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NgModule } from '@angular/core';
 import { AppComponent } from './app.component';
+import { AppRoutesModule } from './app.routing';
 
 @NgModule({
   imports: [
     BrowserModule,
     BrowserAnimationsModule,
+    AppRoutesModule,
   ],
   declarations: [AppComponent],
   bootstrap: [AppComponent],