+++ src/app/modules/shade-demo/shade-demo.module.ts
@@ -1,0 +1,17 @@
+import { NgxLibsModule } from 'ngx-libs';
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+import { ShadeDemoComponent } from './shade-demo.component';
+import { ShadeDemoRoutesModule } from './shade-demo.routing';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    FormsModule,
+    ShadeDemoRoutesModule,
+    NgxLibsModule,
+  ],
+  declarations: [ShadeDemoComponent]
+})
+export class ShadeDemoModule { }