+++ src/app/modules/shade-demo/shade-demo.routing.ts
@@ -1,0 +1,15 @@
+import { ShadeDemoComponent } from './shade-demo.component';
+import { Routes, RouterModule } from '@angular/router';
+import { NgModule } from '@angular/core';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: ShadeDemoComponent,
+  },
+];
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class ShadeDemoRoutesModule { }