+++ projects/ngx-libs/changelog.md
@@ -2,3 +2,4 @@
 ## 2018-12-09
 ### ngx-libs
 - 初始化ngx-libs（wangjx）；
+- 增加shade遮罩组件（wangjx）；